requirements.txt
appdirs==1.4.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==24.2
pyshark==0.6
requests==2.32.3
termcolor==2.5.0
urllib3==2.3.0
Werkzeug==3.1.3
Flask
mysql-connector-python